app/pyproject.toml
[tool.poetry]
name = "stock-analysis"
version = "0.1.0"
description = "Stock market analysis backend"
authors = ["Devin <devin@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
python-dotenv = "^1.0.0"
yfinance = "^0.2.51"
pydantic = "^2.10.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
